requirements.txt
delta-spark==2.3.0
findspark==2.0.1
pyspark==3.3.0
kafka-python==2.0.2
python-dotenv==1.0.0
cassandra-driver==3.28.0
pandas==1.5.3
matplotlib==3.7.2
statsmodels==0.14.0
confluent-kafka
numpy
scikit-learn
keras
backtrader
seaborn
fastapi 
uvicorn
websockets
aiohttp
streamlit
streamlit-lottie
websocket-client
plotly
streamlit_autorefresh
```